+++ client/src/game/api/emits/shape/options.ts
@@ -8,6 +8,7 @@ import { wrapSocket } from "../../helpers";
 
 export const sendShapeSetInvisible = wrapSocket<ShapeSetBooleanValue>("Shape.Options.Invisible.Set");
 export const sendShapeSetDefeated = wrapSocket<ShapeSetBooleanValue>("Shape.Options.Defeated.Set");
+export const sendShapeSetOddHexOrientation = wrapSocket<ShapeSetBooleanValue>("Shape.Options.OddHexOrientation.Set");
 export const sendShapeSetLocked = wrapSocket<ShapeSetBooleanValue>("Shape.Options.Locked.Set");
 export const sendShapeSetIsToken = wrapSocket<ShapeSetBooleanValue>("Shape.Options.Token.Set");
 export const sendShapeSetBlocksMovement = wrapSocket<ShapeSetBooleanValue>("Shape.Options.MovementBlock.Set");